Improve problem #12
Only search dividers up to the square of the number.

See explanation: https://stackoverflow.com/questions/6800193/what-is-the-most-efficient-way-of-finding-all-the-factors-of-a-number-in-python

"sqrt(x) * sqrt(x) = x. So if the two factors are the same, they're both the
square root. If you make one factor bigger, you have to make the other factor
smaller. This means that one of the two will always be less than or equal to
sqrt(x), so you only have to search up to that point to find one of the two
matching factors. You can then use x / fac1 to get fac2."